requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cachetools==5.3.0
diff-match-patch==20200713
Django==4.1.5
django-allauth==0.52.0
django-filter==22.1
django-gsheets==0.0.10
django-import-export==3.0.2
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
Markdown==3.4.1
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.7.1
requests-oauthlib==1.3.1
rsa==4.9
sqlparse==0.4.3
tablib==3.3.0
uritemplate==4.1.1
xlrd==2.0.1
xlwt==1.3.0
